+++ test/Fossa/API/TypesSpec.hs
@@ -54,6 +54,9 @@ genIssue =
     <*> genIssueType
     <*> Gen.maybe genIssueRule
     <*> Gen.maybe arbitraryText
+    <*> Gen.maybe arbitraryText
+    <*> Gen.maybe arbitraryText
+    <*> Gen.maybe arbitraryText
 
 genIssueType :: Gen IssueType
 genIssueType =